// modules/audit-logging/domain/audit-event-type.ts
export enum AuditEventType {
  LOGIN_SUCCEEDED = "LOGIN_SUCCEEDED",
  LOGIN_FAILED = "LOGIN_FAILED",
  MODULE_ACCESSED = "MODULE_ACCESSED",
  FEATURE_ACCESSED = "FEATURE_ACCESSED",
  USER_BLOCKED = "USER_BLOCKED",
  USER_UNBLOCKED = "USER_UNBLOCKED",
  INVOICE_CREATED = "INVOICE_CREATED",
  INVOICE_UPDATED = "INVOICE_UPDATED",
  FINANCIAL_EVENT = "FINANCIAL_EVENT",
}